requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.2
dj-database-url==2.2.0
Django==5.0.6
django-crispy-forms==2.1
django-environ==0.11.2
fontawesomefree==6.5.1
gunicorn==22.0.0
packaging==24.0
psycopg2-binary==2.9.9
sqlparse==0.5.0
typing_extensions==4.12.1
whitenoise==6.6.0
